pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "market_research_agent"
version = "0.1.0"
authors = [{ name = "Market Research Team" }]
description = "Factor-scoring quantitative agent for Indian equity markets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.3.0",
    "yfinance>=0.2.28",  # Yahoo Finance API
    "pandas-ta>=0.3.14b",  # Technical Analysis
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0.1",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",  # Async PostgreSQL
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "python-jose[cryptography]>=3.3.0",  # JWT
    "xgboost>=2.0.0",
    "shap>=0.43.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.4",
]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.291",
]

[tool.setuptools.packages.find]
include = ["market_research_agent*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=market_research_agent --cov-report=term-missing"
asyncio_mode = "auto"